Conflicts:
	Gemfile
	Gemfile.lock

+++ lib/services/ddns_expire.rb
@@ -32,7 +32,7 @@ module Services
         intersection_account_mac_address_and_serial_number << username
       end
 
-      warning_user_emails = Array.new
+      @log_array = Array.new
       intersection_account_mac_address_and_serial_number.each do |i|
         device = Device.find_by(mac_address: i[0], serial_number: i[1])
 
@@ -47,10 +47,18 @@ module Services
           expire_days = (current_time - xmpp_last_username.last_signout_at)/(24*60*60)
 
           DdnsMailer.notify_comment(user, device, xmpp_last_username).deliver_now
-          @rake_log.info "  sent mail to DDNS: #{ device.ddns.hostname }.#{device.ddns.domain.domain_name} expired #{ expire_days } days of Device: #{ device.serial_number } of User: #{ user.email }"
+          # @rake_log.info "  sent mail to DDNS: #{ device.ddns.hostname }.#{device.ddns.domain.domain_name} expired #{ expire_days } days of Device: #{ device.serial_number } of User: #{ user.email }"
 
+          info = {
+            :user => user.email,
+            :ddns => "#{ device.ddns.hostname }.#{device.ddns.domain.domain_name}",
+            :expire_days => expire_days,
+            :device => device.serial_number
+          }
+          @log_array << info
         end
       end
+      @log_array
     end
 
     def self.delete
@@ -80,24 +88,48 @@ module Services
         intersection_account_mac_address_and_serial_number << username
       end
 
-      warning_user_emails = Array.new
+      @log_array = Array.new
       intersection_account_mac_address_and_serial_number.each do |i|
         device = Device.find_by(mac_address: i[0], serial_number: i[1])
 
         if device.present? && device.pairing.present? && device.pairing.first.user.present?
-          warning_user_emails << device.pairing.first.user.email
 
-          device.ddns.destroy
-          delete_route53_record(device.ddns)
+
+          route53, info, result, error = delete_route53_record(device.ddns)
+          if result == "delete route53 succeed"
+            device.ddns.destroy
+            if device.ddns.destroyed?
+              @result_db = "delete ddns db succeed"
+            else
+              @result_db = "delete ddns db failed"
+            end
+          else
+            @result_db = "delete ddns db failed"
+          end
 
           user = device.pairing.first.user
           xmpp_last_username = XmppLast.find_by_decive(device)
           expire_days = (current_time - xmpp_last_username.last_signout_at)/(24*60*60)
 
-          @rake_log.info "  delete DDNS: #{ device.ddns.hostname }.#{device.ddns.domain.domain_name} expired #{ expire_days } days of Device: #{ device.serial_number } of User: #{ user.email }"
+          # @rake_log.info "  delete DDNS: #{ device.ddns.hostname }.#{device.ddns.domain.domain_name} expired #{ expire_days } days of Device: #{ device.serial_number } of User: #{ user.email }"
 
+          info = {
+            :user => user.email,
+            :ddns => "#{ device.ddns.hostname }.#{device.ddns.domain.domain_name}",
+            :expire_days => expire_days,
+            :device => device.serial_number,
+            :result_db => @result_db,
+            :route53 => {
+              :route53_record => route53,
+              :info => info,
+              :result => result,
+              :error => error
+            }
+          }
+          @log_array << info
         end
       end
+      @log_array
     end
 
     def self.create_route53_record(ddns)
@@ -112,11 +144,14 @@ module Services
 
       begin
         rrset = rrsets.create(target_name, 'A', ttl: 300, resource_records: [{value: ddns.get_ip_addr}])
-        @rake_log.info "  Create DDNS record: #{rrset.name}"
-        @rake_log.info "                  ip: #{rrset.resource_records.first[:value]}"
+        rrset_name = rrset.name if rrset
+        rrset_ip = rrset.resource_records.first[:value] if rrset
+        # @rake_log.info "  Create DDNS record: #{rrset.name}"
+        # @rake_log.info "                  ip: #{rrset.resource_records.first[:value]}"
       rescue Exception => error
-        @rake_log.error error
+        # @rake_log.error error
       end
+      return rrset_name, rrset_ip, error
     end
 
     def self.delete_route53_record(ddns)
@@ -131,11 +166,19 @@ module Services
 
       begin
         rrset = rrsets[target_name, 'A']
+        rrset_name = rrset.name if rrset
         info = rrset.delete
-        @rake_log.info "  Delete DDNS record: #{rrset.name}, status: #{info.status}"
+        info_status = info.status if info
+        if rrset.exists? == false
+          result = "delete route53 succeed"
+        else
+          result = "delete route53 failed"
+        end
+        # @rake_log.info "  Delete DDNS record: #{rrset.name}, status: #{info.status}"
       rescue Exception => error
-        @rake_log.error error
+        # @rake_log.error error
       end
+      return rrset_name, info_status, result, error
     end
 
   end
